+++ example/a.ts
@@ -1,12 +1,11 @@
 import { Elysia, t } from '../src'
 
-const route = new Elysia({ aot: false })
-	.get('/valid', () => ({ foo: 'a' }), {
-		afterHandle: () => ({ q: 'a' }),
-		response: t.Object({
-			foo: t.String()
+new Elysia()
+	.model({
+		ids: t.Object({
+			ids: t.Array(t.Union([t.String(), t.ArrayString()]))
 		})
 	})
-	.listen(3000)
-
-// console.log(route.routes[0].compile().toString())
+	.get('/', ({ query }) => query, {
+		query: 'ids'
+	})