+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Bug fix:
 - [#859](https://github.com/elysiajs/elysia/issues/859) clean non-root additionalProperties
 - [#924](https://github.com/elysiajs/elysia/issues/924) decode path param
 - [#985](https://github.com/elysiajs/elysia/issues/924) Nullable accept options
+- [#1028](https://github.com/elysiajs/elysia/issues/1028) string | string[] query parameter, reference array
 - [#1274](https://github.com/elysiajs/elysia/issues/1274) support .use(undefined | false) for conditional plugin
 - [#1276](https://github.com/elysiajs/elysia/issues/1276) mapResponse with set inference produce invalid instruction
 - [#1268](https://github.com/elysiajs/elysia/issues/1268) using number instead of stringifed value for reporting validation error